fix: use u64

+++ src/rollups/nitro/nitro.rs
@@ -175,22 +175,22 @@ impl Rollup for Nitro {
             return VerificationResult::failure();
         }
 
-        let start_message_position = context.next_batch_start_pos as u32;
-        let end_message_position = start_message_position + batch_messages.len() as u32 - 1;
+        let start_message_position = context.next_batch_start_pos;
+        let end_message_position = start_message_position + batch_messages.len() as u64 - 1;
         let start_espresso_block = queue[0..batch_messages.len()]
             .iter()
             .map(|e| e.hotshot_height())
             .min()
-            .unwrap_or(0) as u32;
+            .unwrap_or(0);
         let after_delayed_messages_read = queue[batch_messages.len() - 1]
             .feed_message
             .message
-            .delayed_messages_read as u32;
+            .delayed_messages_read;
         let min_espresso_block_still_in_queue = queue[batch_messages.len()..]
             .iter()
             .map(|e| e.hotshot_height())
             .min()
-            .unwrap_or(0) as u32;
+            .unwrap_or(0);
         VerificationResult {
             success: true,
             start_message_position,
